feat: implement Phase 9 Agent DX improvements (tasks 9.1-9.4)
Task 9.1 - Configurable Access Token Lifetime per Credential:
- Add token_lifetime_seconds column to Credential model
- Add token_lifetime_seconds field to CredentialCreate (ge=60, le=86400)
- Surface token_lifetime_seconds in CredentialResponse
- Use custom lifetime when minting tokens via client_credentials grant
- Alembic migration: add_token_lifetime_seconds_to_credentials

Task 9.2 - Surface Quickstart Scope Constraints via Discover and Errors:
- Add 'quickstart' key to /.well-known/agent-configuration with description,
  scopes:[], and next_steps
- Add 'hint' field to invalid_scope error when credential has empty scopes

Task 9.3 - Make Delegation Chain Depth Semantics Explicit:
- Add chain_depth_remaining to DelegationResponse (max_chain_depth - chain_depth)
- Add max_absolute_chain_depth alias to DelegationResponse
- Add 'delegation' section to /discover with depth_semantics and description
- Improve chain-depth-exceeded error with structured detail fields

Task 9.4 - Emit Token Expiry Warning via Webhook and Response Header:
- Add TOKEN_EXPIRY_WARNING_SECONDS config (default 120)
- Add X-Token-Expires-In header on all authenticated responses
- Add X-Token-Refresh-Advised: true when remaining lifetime < threshold
- Emit token.expiring_soon webhook event (Redis-deduped per family_id)
- Emit token.refreshed webhook event after successful refresh grant

Tests: 357 unit tests passing

+++ src/agentauth/services/token.py
@@ -38,7 +38,7 @@ class TokenService:
         scopes: list[str] | None = None,
         audience: str | list[str] | None = None,
         token_type: str = "access",
-        expires_in_minutes: int | None = None,
+        expires_in_minutes: float | None = None,
         delegation_chain: list[UUID] | None = None,
         algorithm: KeyAlgorithm = KeyAlgorithm.RS256,
         family_id: str | None = None,
